TST Remove serpentTools.tests; External test module
The tests/utils.py file has been merged into tests/__init__.py
and should be used for all testing. Imports of the form
from serpentTools.tests foo
have been replaced with from tests import foo.

Helper function testPlotAttrs has been renamed as plotAttrTest
so that is is missed with pytest
