+++ setup.py
@@ -47,7 +47,7 @@ pythonRequires = '>=2.7,!=3.0,!=3.1,!=3.2,!=3.3,!=3.4'
 setupArgs = {
     'name': 'serpentTools',
     'packages': ['serpentTools', 'serpentTools.parsers', 'serpentTools.utils',
-                 'serpentTools.data', 'serpentTools.tests', 'serpentTools.io',
+                 'serpentTools.data', 'serpentTools.io',
                  'serpentTools.objects', 'serpentTools.samplers'],
     'url': 'https://github.com/CORE-GATECH-GROUP/serpent-tools',
     'description': ('A suite of parsers designed to make interacting with '
@@ -74,7 +74,6 @@ if HAS_SETUPTOOLS:
     setupArgs.update({
         'python_requires': pythonRequires,
         'install_requires': installRequires,
-        'test_suite': 'serpentTools.tests'
     })
 
 setup(**setupArgs)
