+++ tests/test_detSampler.py
@@ -22,15 +22,14 @@ File Descriptions
 
 """
 from six import iteritems
-
 from numpy import square, sqrt
 from numpy.testing import assert_allclose
-
 from serpentTools.messages import MismatchedContainersError
 from serpentTools.data import getFile
 from serpentTools.parsers.detector import DetectorReader
 from serpentTools.samplers.detector import DetectorSampler
-from serpentTools.tests.utils import TestCaseWithLogCapture
+
+from tests import TestCaseWithLogCapture
 
 _DET_FILES = {
     'bwr0': 'bwr_0',
